# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "智能学术界人才库 API",
    "description": "智能人才库 V2.0.0 API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/collect/tasks": {
      "get": {
        "tags": [
          "Collect Configuration"
        ],
        "summary": "获取采集任务列表",
        "description": "获取采集任务列表（分页）",
        "operationId": "list_tasks_api_v1_collect_tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按状态筛选",
              "title": "Status"
            },
            "description": "按状态筛选"
          },
          {
            "name": "tech_element_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按技术要素筛选",
              "title": "Tech Element Id"
            },
            "description": "按技术要素筛选"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectTaskListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Collect Configuration"
        ],
        "summary": "触发采集任务",
        "description": "触发采集任务。\n\n**固定参数：**\n- 数据类型：学者、论文、机构\n- 时间范围：2010.1.1 至今\n\n**可配置参数：**\n- tech_element_id：技术要素ID\n- collect_mode：full（全量）或 incremental（增量）\n\n**说明：**\n采集任务在后台异步执行，可通过任务列表查看进度。",
        "operationId": "trigger_task_api_v1_collect_tasks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerCollectTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collect/tasks/{task_id}": {
      "get": {
        "tags": [
          "Collect Configuration"
        ],
        "summary": "获取采集任务详情",
        "description": "获取指定采集任务的详细信息，包含执行日志",
        "operationId": "get_task_api_v1_collect_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Collect Configuration"
        ],
        "summary": "删除采集任务",
        "description": "删除已完成的采集任务记录",
        "operationId": "delete_task_api_v1_collect_tasks__task_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collect/tasks/{task_id}/execute": {
      "post": {
        "tags": [
          "Collect Configuration"
        ],
        "summary": "执行采集任务",
        "description": "执行待执行状态的采集任务",
        "operationId": "execute_task_api_v1_collect_tasks__task_id__execute_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collect/tasks/{task_id}/cancel": {
      "post": {
        "tags": [
          "Collect Configuration"
        ],
        "summary": "取消采集任务",
        "description": "取消正在执行的采集任务",
        "operationId": "cancel_task_api_v1_collect_tasks__task_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collect/tasks/active": {
      "get": {
        "tags": [
          "Collect Configuration"
        ],
        "summary": "获取活动任务",
        "description": "获取当前正在执行或待执行的任务",
        "operationId": "get_active_tasks_api_v1_collect_tasks_active_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CollectTaskResponse"
                  },
                  "type": "array",
                  "title": "Response Get Active Tasks Api V1 Collect Tasks Active Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/collect/tasks/{task_id}/progress": {
      "get": {
        "tags": [
          "Collect Configuration"
        ],
        "summary": "获取任务详细进度",
        "description": "获取任务的详细进度信息，包括各Venue子任务状态",
        "operationId": "get_task_detailed_progress_api_v1_collect_tasks__task_id__progress_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CollectTaskListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CollectTaskResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "CollectTaskListResponse",
        "description": "采集任务列表响应"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}